Store CQL in the URL, not JSON
